+++ src/main/java/roomescape/controller/api/ReservationApiController.java
@@ -13,8 +13,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.AuthenticatedMember;
 import roomescape.controller.api.validator.IdPositive;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationWaitWithRank;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservationwait.ReservationWaitWithRank;
 import roomescape.domain.member.Member;
 import roomescape.service.dto.request.ReservationSaveRequest;
 import roomescape.service.dto.response.reservation.ReservationResponse;
