+++ src/main/java/roomescape/controller/api/admin/AdminReservationWaitApiController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.ReservationWait;
+import roomescape.domain.reservationwait.ReservationWait;
 import roomescape.service.dto.response.reservationwait.ReservationWaitResponses;
 import roomescape.service.reservationwait.ReservationWaitDeleteService;
 import roomescape.service.reservationwait.ReservationWaitFindService;
