+++ src/main/java/roomescape/domain/reservationwait/ReservationWait.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservationwait;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -9,6 +9,8 @@ import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
 import java.time.LocalDate;
 import java.util.Objects;
+import roomescape.domain.ReservationTime;
+import roomescape.domain.Theme;
 import roomescape.domain.member.Member;
 
 @Entity
