+++ src/main/java/roomescape/domain/reservationwait/ReservationWaitWithRank.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservationwait;
 
 public class ReservationWaitWithRank {
