+++ src/main/java/roomescape/service/dto/request/ReservationWaitSaveRequest.java
@@ -3,8 +3,8 @@ package roomescape.service.dto.request;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 import roomescape.domain.ReservationTime;
-import roomescape.domain.ReservationWait;
-import roomescape.domain.ReservationWaitStatus;
+import roomescape.domain.reservationwait.ReservationWait;
+import roomescape.domain.reservationwait.ReservationWaitStatus;
 import roomescape.domain.Theme;
 import roomescape.domain.member.Member;
