+++ src/main/java/roomescape/service/dto/response/reservation/ReservationResponse.java
@@ -1,7 +1,7 @@
 package roomescape.service.dto.response.reservation;
 
 import java.time.LocalDate;
-import roomescape.domain.Reservation;
+import roomescape.domain.reservation.Reservation;
 import roomescape.service.dto.response.member.MemberIdAndNameResponse;
 import roomescape.service.dto.response.reservationTime.ReservationTimeResponse;
 import roomescape.service.dto.response.theme.ThemeResponse;
