+++ src/main/java/roomescape/service/dto/response/reservation/ReservationResponses.java
@@ -1,7 +1,7 @@
 package roomescape.service.dto.response.reservation;
 
 import java.util.List;
-import roomescape.domain.Reservation;
+import roomescape.domain.reservation.Reservation;
 
 public record ReservationResponses(List<ReservationResponse> reservations) {
