+++ src/main/java/roomescape/service/reservationtime/ReservationTimeFindService.java
@@ -3,8 +3,8 @@ package roomescape.service.reservationtime;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationStatuses;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservation.ReservationStatuses;
 import roomescape.domain.ReservationTime;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
