+++ src/main/java/roomescape/service/reservationwait/ReservationWaitDeleteService.java
@@ -2,7 +2,7 @@ package roomescape.service.reservationwait;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.ReservationWait;
+import roomescape.domain.reservationwait.ReservationWait;
 import roomescape.exception.InvalidRequestException;
 import roomescape.repository.ReservationWaitRepository;
