+++ src/test/java/roomescape/service/reservation/ReservationDeleteTest.java
@@ -2,8 +2,8 @@ package roomescape.service.reservation;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static roomescape.domain.ReservationWaitStatus.CONFIRMED;
-import static roomescape.domain.ReservationWaitStatus.WAITING;
+import static roomescape.domain.reservationwait.ReservationWaitStatus.CONFIRMED;
+import static roomescape.domain.reservationwait.ReservationWaitStatus.WAITING;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -11,9 +11,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import roomescape.domain.Reservation;
+import roomescape.domain.reservation.Reservation;
 import roomescape.domain.ReservationTime;
-import roomescape.domain.ReservationWait;
+import roomescape.domain.reservationwait.ReservationWait;
 import roomescape.domain.Theme;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberEmail;
