+++ src/test/java/roomescape/service/reservationwait/ReservationWaitDeleteServiceTest.java
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import roomescape.domain.ReservationTime;
-import roomescape.domain.ReservationWait;
-import roomescape.domain.ReservationWaitStatus;
+import roomescape.domain.reservationwait.ReservationWait;
+import roomescape.domain.reservationwait.ReservationWaitStatus;
 import roomescape.domain.Theme;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberEmail;
